+++ src/index.ts
@@ -1,5 +1,6 @@
-import { DataApi, FileMakerError } from "./client.js";
+import { DataApi, FileMakerError, isOttoAuth } from "./client.js";
 
-export { FileMakerError, DataApi };
-export { removeFMTableNames } from "./utils/utils.js";
+export { DataApi, FileMakerError, isOttoAuth };
+export * from "./utils/utils.js";
+
 export default DataApi;