+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 import { generateSchemas } from "./codegen.js";
 import { removeFMTableNames } from "./utils.js";
+export { isOtto3APIKey, isOttoAPIKey, isOttoFMSAPIKey } from "./utils.js";
 
 export { removeFMTableNames, generateSchemas };